+++ CHANGELOG.md
@@ -4,16 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## \[Unreleased]
+
+## \[0.2.0] - 2017-03-28
 ### Added
 -   Course model.
 -   REST API for courses.
 -   Server-side Session.
--   OAuth2 Authentication with Mesa.
+-   OAuth2 Authentication with Course Dev.
 
 ### Changed
 -   CHANGELOG.md format.
 
-## [0.1.0] - 2017-02-27
+## \[0.1.0] - 2017-02-27
 ### Added
 -   Project skeleton.